new updates, 09/16/2019
update for reduced turning speed while arm is above a certain height, updated to include heights for cargo.  keep in mind that cargo heights will have to be calibrated later once we get the exact rocket model for cargo together.  the formula part of lowering speed was removed because it caused some serious issues, such as the arm slamming back into the robot.  eventually this will be readjusted.