Robot.java Competition Grade Alpha 0.8.9
Now includes features such as the "PI" parts of the PID, slowing down as it approaches its target.  Has slowed turning while arm is raised, the switch on the second controller now controls the speed of the intake wheels, etc.  We've tested the current arm raising and lowering feature, the slow turning while arm raised, and the intake wheel speed.  More updates to come based on more testing.